fix warnings

+++ bigint_tiny.h
@@ -128,7 +128,7 @@ struct BigIntTiny {
         double db = b.v.back() + (b.v.size() > 1 ? b.v[b.v.size() - 2] / 1e4 : 0) + (b.v.size() > 2 ? (b.v[b.v.size() - 3] + 1) / 1e8 : 0);
         for (int i = (int)v.size() - 1; i >= 0; i--) {
             c.v.insert(c.v.begin(), v[i]);
-            int m = (int)((c.get_pos(b.v.size()) * 10000 + c.get_pos(b.v.size() - 1)) / db);
+            int m = (int)((c.get_pos((int)b.v.size()) * 10000 + c.get_pos((int)b.v.size() - 1)) / db);
             c = c - b * m, d.v[i] += m;
             while (!(c < b))
                 c = c - b, d.v[i] += 1;
